backend/requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.73.0
anyio==4.11.0
attrs==25.4.0
certifi==2025.11.12
click==8.3.0
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.121.2
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
jiter==0.12.0
multidict==6.7.0
packaging==25.0
pillow==12.0.0
pluggy==1.6.0
propcache==0.4.1
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.1
pytest-asyncio==1.3.0
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
redis==7.0.1
replicate==1.0.7
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.3
structlog==25.5.0
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
